New Changes
Fixed the bug that was not adding numbers. The bug was the value we stored inside of the array was a string because the button we clicked came from an array which had string as an number. We also made new changes to the code we can now do arithmetic. Next add better features like displaying the result and other